moved models to a separate folder
