+++ TinkoffStocks/Source/Services/CandlesRepository/Models/CandleStickInterval.swift
@@ -1,0 +1,17 @@
+//
+//  CandleStickInterval.swift
+//  T-Stocks
+//
+//  Created by sleepcha on 1/15/25.
+//
+
+enum CandleStickInterval: CaseIterable {
+    case min5
+    case min15
+    case min30
+    case hour1
+    case hour4
+    case day
+    case week
+    case month
+}
